Use panamized version for windows in Int7VectorScorer
