+++ libs/simdvec/src/test/java/org/elasticsearch/simdvec/internal/vectorization/ES92Int7VectorScorerTests.java
@@ -27,6 +27,15 @@ import static org.hamcrest.Matchers.greaterThan;
 
 public class ES92Int7VectorScorerTests extends BaseVectorizationTests {
 
+    public boolean hasNativeAccess() {
+        var jdkVersion = Runtime.version().feature();
+        var arch = System.getProperty("os.arch");
+        var osName = System.getProperty("os.name");
+        return (jdkVersion >= 21
+            && (arch.equals("aarch64") && (osName.startsWith("Mac") || osName.equals("Linux"))
+                || arch.equals("amd64") && osName.equals("Linux")));
+    }
+
     public void testInt7DotProduct() throws Exception {
         // only even dimensions are supported
         final int dimensions = random().nextInt(1, 1000) * 2;
@@ -52,7 +61,9 @@ public class ES92Int7VectorScorerTests extends BaseVectorizationTests {
                 final IndexInput slice = in.slice("test", 0, (long) dimensions * numVectors);
                 final IndexInput slice2 = in.slice("test2", 0, (long) dimensions * numVectors);
                 final ES92Int7VectorsScorer defaultScorer = defaultProvider().newES92Int7VectorsScorer(slice, dimensions);
+                assertFalse(defaultScorer.hasNativeAccess());
                 final ES92Int7VectorsScorer panamaScorer = maybePanamaProvider().newES92Int7VectorsScorer(slice2, dimensions);
+                assertEquals(panamaScorer.hasNativeAccess(), hasNativeAccess());
                 for (int i = 0; i < numVectors; i++) {
                     in.readBytes(vector, 0, dimensions);
                     long val = VectorUtil.dotProduct(vector, query);
@@ -119,7 +130,9 @@ public class ES92Int7VectorScorerTests extends BaseVectorizationTests {
                 // padding bytes.
                 final IndexInput slice = in.slice("test", 0, (long) (dimensions + 16) * numVectors);
                 final ES92Int7VectorsScorer defaultScorer = defaultProvider().newES92Int7VectorsScorer(in, dimensions);
+                assertFalse(defaultScorer.hasNativeAccess());
                 final ES92Int7VectorsScorer panamaScorer = maybePanamaProvider().newES92Int7VectorsScorer(slice, dimensions);
+                assertEquals(panamaScorer.hasNativeAccess(), hasNativeAccess());
                 for (int i = 0; i < numVectors; i++) {
                     float scoreDefault = defaultScorer.score(
                         qQuery,
@@ -198,7 +211,9 @@ public class ES92Int7VectorScorerTests extends BaseVectorizationTests {
                 // padding bytes.
                 final IndexInput slice = in.slice("test", 0, (long) (dimensions + 16) * numVectors);
                 final ES92Int7VectorsScorer defaultScorer = defaultProvider().newES92Int7VectorsScorer(in, dimensions);
+                assertFalse(defaultScorer.hasNativeAccess());
                 final ES92Int7VectorsScorer panamaScorer = maybePanamaProvider().newES92Int7VectorsScorer(slice, dimensions);
+                assertEquals(panamaScorer.hasNativeAccess(), hasNativeAccess());
                 float[] scoresDefault = new float[ES91Int4VectorsScorer.BULK_SIZE];
                 float[] scoresPanama = new float[ES91Int4VectorsScorer.BULK_SIZE];
                 for (int i = 0; i < numVectors; i += ES91Int4VectorsScorer.BULK_SIZE) {
